backend/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
python-multipart
numpy
scipy
clip @ git+https://github.com/openai/CLIP.git